requirements-test.txt
pylint
build
twine
pytest
pytest-asyncio
pre-commit
mypy
ruff
pytest_httpx
